+++ duckscript_sdk/src/sdk/std/env/mod.rs
@@ -1,6 +1,7 @@
 mod get;
 mod get_home_dir;
 mod get_user_name;
+mod is_windows;
 mod os_family;
 mod print_current_directory;
 mod set;
@@ -19,6 +20,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(get::create(&package))?;
     commands.set(get_home_dir::create(&package))?;
     commands.set(get_user_name::create(PACKAGE))?;
+    commands.set(is_windows::create(PACKAGE)?)?;
     commands.set(os_family::create(PACKAGE))?;
     commands.set(print_current_directory::create(&package))?;
     commands.set(set::create(&package))?;